Refactor: Transform project into an Arduino library
This commit refactors the entire project from a standalone application into a formal Arduino library named `xDuinoRails_MotorDriver`.

Key changes include:
- All motor control logic is encapsulated within the `XDuinoRails_MotorDriver` class.
- The original application has been recreated as an example sketch in `examples/Demo/Demo.ino`.
- The project is now correctly structured with a dedicated `platformio.ini` for the example.
- A `src/native_main.cpp` file has been added to provide a `main()` function for the `native` environment, resolving the CI build failure. This file is excluded from Arduino builds.

**CI Note:** To build the example firmware, the CI command must be updated to: `pio run --project-dir examples/Demo`
